requirements.txt
# --- CORE FRAMEWORK ---
PySide6>=6.5.0       # UI Framework
pydantic>=2.0.0      # Data models (v2 is required for .model_dump in utils.py)
typing-extensions    # Backwards compatibility for type hints

# --- AI & LLM ---
ollama               # Local LLM communication
aiohttp              # Async communication for GUI responsiveness
python-dotenv        # Environment/Settings management

# --- DOCUMENT PARSING ---
pdfplumber           # PDF layout extraction
python-docx          # Word doc extraction
striprtf             # RTF cleaning

# --- DATA & EXPORTS ---
pandas               # CSV/Excel logic
openpyxl             # Excel engine
lxml                 # Advanced XML for Movie Magic (.sex) export